wmf-sunset-patch: add EntityTag.getdetails api
This is an added entity used by WMF modules. It needs review as I think
it is not really aligned with the standard api at the moment

The issue is that it is retrieving labels without a second query
by adding a join - which is superior to chaining. However, I suspect the api
way of using pseudoconstants would be equally performant.

So, keeping this for 4.6 but tagging for later review & not opening a JIRA
at this stage as I'm not sure there is an upstream issue.

It is also possible to add api actions in a module or extension if we
can't find a better option

entity-tag

fix tag details