CRM-17641 backport Civi class in order to support other backports
This class supports the reduction of the infamous utf names query (Bug: T116724 ) but does nothing of itself.

It is of interest from the point of view of deployment because
1) I had to run the deployment script we have been working on to update composer - which we have as a checked in library and
2) The deployment script had to have an upstream file backported before it would include the Civi.php file which was not in it's
original whitelist. However, this doesn't make me feel better about using a blacklist (a list of files to delete rather than a list of
files to keep/copy as I'd rather miss a file & have something break than miss a file & deploy an insecure file never intended to go
on a prod site

It's worth noting that I agreed with Tim that this backport would go into the next 4.6 release (but we will drag our heels on the utf names one a big).

Bug: T116724

Change-Id: I7c8c4c5d556b7a1bf820b744f9c0dda4359ab6c8