(patch already in 4.7) CRM-17640 Update contribution api to do a sensible query when getcount is being invoked
Without this patch contribution.getcount api does a full query with the full set of select fields and
the nasty joins those entail and only after doing that does it filter for count.

This patch re-uses the code in use for contact api query for contribution getcount

The slow query is experienced when editing to refund.

Bug: T116317
Change-Id: I5e3767b8828ac8c72d2279a66812c947dfe66e17