CRM-17837 Improve Lybunt report.
The WMF Lybunt report was only working for calendar year and not fiscal year which was blocking closing some tickets:

Bug: T119911 (sort in descending order of last year's donation amount)
Bug: T119909 (show last year total in the stats at the bottom of the report)
Bug: T122027 (format custom money columns as money)

We also had two open performance related bugs on Lybunt:

Bug: T119036 (calendar year query is slow - I think it meant fiscal year is slow).
Bug: T122024 (speed up lybunt - this was raised for one specific aspect of the query - ie. the use of the YEAR function rather than BETWEEN in the date selection clause)

On digging I could not see how the the existing methodology could be used to make the Fiscal year LYBUNT work without adding a whole bunch more fields (alluded to in T119036).

However, my tests didn't indicate that using these custom fields gave more speed improvement than fixing up the core lybunt report query to be more efficient. I've added notes about my query digging here https://issues.civicrm.org/jira/browse/CRM-17837.

 Having made the changes described on the CRM-17837 core ticket I'm getting the results for either fiscal year or calendar year in around 30 seconds.

That still isn't 'snappy' but the WMF lybunt is currently taking around 4 minutes. The main filter query is 13 seconds, which the smart group filter takes around 10 seconds. I might be able to shave a bit more off it but we are getting to the point where it is fairly acceptable.

 It would be nice to tackle the smart group filter at some point and I would possibly prioritise that if we are to try to get more speed.

I tried running the queries without any smart group filter out of interest and to get ALL the people who gave in 14/15 but not 15/16 it took around 8 minutes to do the main queries - although the web Ui had long timed out by the time they completed.

I was current thinking is that assuming there is no issue found in core QA and fact an extra field is added (contact sort_name as a required field) are OK I think we should decommission the custom WMF report.

However, I identified a gap in the core report in that it does not include soft credit payments in the calculations.

The comments after this are comments I added originally but my underlying premise - that the last_donation_date field take soft-credits into account is now something I'm uncertain of and need to re-review.

I've kept these comments as they do add information but they should be understood as interim thinking on the issue:

If The custom field wmf_donor.last_donation_date does take soft credits into account & I can updated the wmf Lynbut to use this - however, the impact is that it then becomes a Lybuns report rather than Lybunt. When viewing the current year there is no real impact but viewing historical results there are:
ie. if I gave in 2013 and 2015 in a Lybunt
 * I would show up on the 2014 and 2016 reports. For a Lybuns I would only show up on the 2016 report
 * because i gave since 2014.
 *
 * If this limitation is significant the options are to
 * 1) accept Lybuns for Fiscal year and make this do a Lybunt for calendar year - we have the
 *   custom fields already for that OR
 * 2) create a new set of custom fields for fiscal years ie is_2014_5_donor.
 * 3) Enhance the core Lybunt report to allow soft credits to be considered when calculating last donation.
I am going to detail all the changes to the core report I made on the JIRA ticket - there are quite a few.
I hope to close all the tickets mentioned above.

Bug: T123671

Change-Id: If33c0a42c699b07d21f48d7bec8c35a261c54d13