Back port changes to support altering indexes in an upgrade-proof way
This is a combination of 2 commits I added to 4.7. The purpose is to add functions to core to allow
indexes to be added or deleted in a way that won't crash the upgrade script if they are already present
/ deleted.

Separately I have submitted patches to core to add 2 indexes and drop 2 indexes in the 4.7.1 upgrade script.

I will separately submit a patch to our wmf_civicrm upgrade script to add the 2 missing indexes using these
functions.

(I'm on the fence about the 2 removals as I have no evidence they are important for performance and
they are mostly for tidy up purposes. I'll leaning towards leaving for now and we'll catch them next upgrade.)

https://issues.civicrm.org/jira/browse/CRM-17775
https://issues.civicrm.org/jira/browse/CRM-17882

Backported patches (tests included in original patch but not in the backport)

CRM-17881 extend createIndex function to create combined indexes
CRM-17775 Add safe function for dropping indexes

Bug: T122947
Change-Id: I23fbd5860aec3b66a294e12d7cb2f07feb27aedc