Search speed fix - partial backport of CRM-17492
This is partial backport of https://github.com/civicrm/civicrm-core/pull/7432/files in order to address slowness on searching on
contribution_source field. There are a couple of minor errors in the upstream version (ie. repetitive check on setting & a typo). This version has the typo fixed but not the other. I didn't want to get drawn in & haven't fixed upstream. Neither are of any real significance.

Bug: T123418
Change-Id: I7d4401579a0e26c0241adfda1bb70cf43e440952