CRM-18212 allow triggers to be managed offline.
This commit adds a setting which when set (in civicrm.settings.php or through the api or drush - not through the UI) allows you to designate the site
as having it's triggers managed offline.

This is good for when it really is not appropriate to give the web use that mysql permission.
ie  when replication is in play because it can allow privilige escalation

https://github.com/civicrm/civicrm-core/pull/8036

Bug: T131224
Change-Id: Iedc8dfca17be38a9ea2b5848bb069e7c84805bf8