CRM-18193 remove arbitrary setting of log_date when not provided.
In order to be able to ignore the log_conn_date where it is not relevant the logging detail reports need to stop making up a log_conn_date where one is not provided. The code that does that was part of the inital report commit and on removing it I could not replicate any issues - ie. accessing the detail reports without criteria worked fine

Change-Id: Ib426eff48bd866d52bfb43efa67b1d3570d1f45a